Fix bug deleting a  person does not delete the visits of the person
This bug is caused by clear the current data. Before this fix, clear will reset persons and visits list and dictionary of Medlloger
to that of a new Medlogger. The reset logic of visits and persons are clear current list and copy from new list, but the reset logic
of dictionary is copy key-value pairs of new dictionary to current dictionary. This means after clear, dictionary still holds old
key-value pairs, hence leading to bug after inserting person with duplicate details after clear.